Treat deletion of non existing entity as success
