+++ app/api/client_resources.py
@@ -7,6 +7,7 @@ from gql.transport.exceptions import TransportQueryError
 from pydantic import BaseModel, ConfigDict, Field
 from pydantic.alias_generators import to_camel
 
+from app.api.exceptions import GraphQLMutationError
 from app.api.protocol import TwingateClientProtocol
 from app.crds import ProtocolPolicy, ResourceSpec
 
@@ -275,5 +276,10 @@ class TwingateResourceAPIs:
             )
 
             return bool(result["ok"])
+        except GraphQLMutationError as gql_err:
+            if "down not exist" in gql_err.error:
+                return True
+
+            raise
         except TransportQueryError:
             return False
